% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_packages_for_parallel.R
\name{compare_packages_for_parallel}
\alias{compare_packages_for_parallel}
\title{Compare packages for parallel computing}
\usage{
compare_packages_for_parallel()
}
\description{
Surprisingly, performance of the packages for parallel computing differs
substantially. This requires a lot of those packages installed.
}
\examples{
compare_packages_for_parallel()
}
