% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_log.R
\name{.source}
\alias{.source}
\title{Source & log}
\usage{
.source(file, echo = F, path = "../logs")
}
\arguments{
\item{file}{the name of R file to be run}

\item{echo}{same as 'echo' in \code{\link{source}}}

\item{path}{log output directory, relative path from the current}
}
\description{
Execute a script while writing a log to keep the history. The log file is
created at the path given as \var{path}. If it is not available, the file
can be found where the R script \var{file} is. Automatically, the execution
date is added into the log file name. Loaded packages in the script are
listed in the log.
}
\seealso{
\code{\link{source}}
}
