% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.NBISsupport-package.R
\docType{package}
\name{NBISsupport-package}
\alias{NBISsupport-package}
\title{NBISsupport : A collection of functions useful for support projects in NBIS}
\description{
This package includes functions that simplify works here and there
}
\author{
Mun-Gwan Hong <\email{mungwan@gmail.com}>
}
\keyword{package}
