% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper1.R
\name{toupper1}
\alias{toupper1}
\title{Upper 1st Character}
\usage{
toupper1(x)
}
\arguments{
\item{x}{a string vector}
}
\value{
a string vector
}
\description{
Capitalize the first character only
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
