% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R_template.R
\name{use.template}
\alias{use.template}
\title{Use a template}
\usage{
use.template(file, extension, template_file)
}
\arguments{
\item{file}{The name of generating file}

\item{extension}{the extension of the `file`}

\item{template_file}{the template for the `file`}
}
\value{
a file created
}
\description{
Use a template
}
\author{
Mun-Gwan Hong, \email{mungwan@gmail.com}
}
